fix deprecations after r25551: move cope Properties from demoshop into framework
git-svn-id: https://svn.exedio.com/cope/cope/trunk@25554 e7d4fc99-c606-0410-b9bf-843393a9eab7

+++ runtime/src/com/exedio/cope/DataField.java
@@ -96,8 +96,8 @@ public final class DataField extends Field<DataField.Value>
 		this.model = type.getModel();
 		final ConnectProperties properties = model.getConnectProperties();
 		column = new BlobColumn(table, name, optional, maximumLength);
-		bufferSizeDefault = min(properties.dataFieldBufferSizeDefault.intValue(), maximumLength);
-		bufferSizeLimit = min(properties.dataFieldBufferSizeLimit.intValue(), maximumLength);
+		bufferSizeDefault = min(properties.dataFieldBufferSizeDefault, maximumLength);
+		bufferSizeLimit = min(properties.dataFieldBufferSizeLimit, maximumLength);
 
 		return column;
 	}
